refactor(configurator): split config update workflows
